Fix Windows build and test cross-platform
