+++ cmd/version.go
@@ -6,11 +6,9 @@ import (
 	"io"
 	"net/http"
 	"os"
-	"os/user"
 	"runtime"
 	"strconv"
 	"strings"
-	"syscall"
 
 	"github.com/spf13/cobra"
 )
@@ -210,40 +208,4 @@ func printUpdateInstructions(tmp, exe string) {
 			fmt.Printf("To update, run:\n  sudo cp %s %s\n  sudo chmod +x %s\n", tmp, exe, exe)
 		}
 	}
-}
-
-func fileWritable(path string) bool {
-	info, err := os.Stat(path)
-	if err != nil {
-		return false
-	}
-	mode := info.Mode().Perm()
-	if runtime.GOOS == "windows" {
-		// On Windows rely on the write bits only.
-		return mode&0222 != 0
-	}
-	st, ok := info.Sys().(*syscall.Stat_t)
-	if !ok {
-		return mode&0222 != 0
-	}
-	uid := os.Getuid()
-	if int(st.Uid) == uid && mode&0200 != 0 {
-		return true
-	}
-	usr, err := user.Current()
-	if err == nil {
-		gids, _ := usr.GroupIds()
-		for _, g := range gids {
-			gid, _ := strconv.Atoi(g)
-			if int(st.Gid) == gid && mode&0020 != 0 {
-				return true
-			}
-		}
-	} else if int(st.Gid) == os.Getgid() && mode&0020 != 0 {
-		return true
-	}
-	if mode&0002 != 0 {
-		return true
-	}
-	return false
 }
